fix: relax requirements on FunC identifiers to `(.*?)`
The regex to approximate actual FunC naming options is gigantic and too
unwieldy to be put here. That's especially true considering that this
will only be used for highlighting and only in a very contained native
function name mapping scenarios.

The similar "include everything" approach was taken previously in
tact.vim:
https://github.com/tact-lang/tact.vim/blob/ddee48224ac89f4aa93f49d4544193ecd6b6948b/syntax/tact.vim#L77-L83